Add files via upload
Using 'checkerboard.png' to calibrate the camera. Before running the code, the picture should be placed in landscape, that is 10*7.
After placing the picture, press the space bar for 20 times to finish the calibration. Then you can find a file named 'calibrate.npz' is generated, which can be used in the future.